Add a telnet synth server
This barebones implementation of telnet can be used to play around with the network.

I don't know python, so the code is somewhat awkward.

It will not work if you send the string in bulk -- it requires the '\n' or '\r\n' to be sent separately afterwards. Windows' telnet complies because it sends each keystroke separately.